SCISOFT NCD: SCI-793 save dataset errors using IErrorDataset interface

+++ uk.ac.diamond.scisoft.ncd/src/uk/ac/diamond/scisoft/ncd/DetectorResponse.java
@@ -46,7 +46,7 @@ public class DetectorResponse {
 		float[] parenterror = (float[]) ConvertUtils.convert(error, float[].class);
 		
 		float[] mydata = new float[parentdata.length];
-		float[] myerror = new float[parenterror.length];
+		double[] myerror = new double[parenterror.length];
 		
 		float[] responseBuffer = response.getData();
